requirements.txt
django==5.1
djangorestframework==3.15.2
django-jsonfield==1.4.1
bcrypt==4.2.0
PyJWT==2.9.0